Compose: simplify WikiErrorView and context handling.
